# perf: index frequent queries