--- a/migrations/028.do.index_unpublished_tasks.sql
+++ b/migrations/028.do.index_unpublished_tasks.sql
@@ -1,0 +1,3 @@
+CREATE INDEX CONCURRENTLY measurements_not_published
+ON measurements (published_as)
+WHERE published_as IS NULL;