--- a/migrations/029.do.index_retrieval_tasks_round_id.sql
+++ b/migrations/029.do.index_retrieval_tasks_round_id.sql
@@ -1,0 +1,1 @@
+CREATE INDEX CONCURRENTLY retrieval_tasks_round_id ON retrieval_tasks (round_id);